Imitate EOperation for obtaining the property descriptors and remove unnecessary methods
We call super.getPropertyDescriptors not with the actual object of type
BehavioredEOperation but with a new EOperation to get all the choices
for EType of operations as with EOperations. The
ETypedElementItemProvider queries the URI of the provided objects and as
this is not the URI of Ecore when providing a BehavioredEOperation, we
do not get all options, such as Ecore's simple data types.