Fixes issue 22
To enable the property view, we recently changed the selection of
graphiti shape elements to eobjects. This, however, broke the context
menu. Thus, I changed it back to provide the selected graphiti shape
elements via the getSelection() method and overwrote the default
adapter-based propertySourceProvider (cf. setPropertySourceProvider())
to check for graphiti elements.