XMOF configuration map: add new constructor to directly use dynamic objects + Java 1.8
To support cases where the XMOF engine is directely provided with
configuration objects (eg. GEMOC Studio), I've added a new constructor
with a boolean "providedAreConfiguration".

If providedAreConfiguration=true, we bypass the creation of all
configuration objects, and we simply fill the map to have each
configuration object point to itself. This ensures that existing code
relying on the configuration map will keep working.