Add executionContext.reset() at construction + throw exception if problem
ExecutionContext is singleton, which means that if the execution is
interrupted by an exception, it will remain in a bad state for future
executions. This can be mitigated by calling reset() each time a
XMOFVirtualMachine is constructed.

When an exception is encountered during execution, listeners were
notified but nothing else was done. Throwing a RuntimeException ensures
that at least the problem cause gets displayed in the console.